Refactored logger and DatabaseLogger
logger is now in logger/index.ts and DatabaseLogger in logger/DatabaseLogger.ts
